[feat][27] 결제 모듈 테스트 코드 작성. 결제 환불 api 추가 구현. 가격 필드 타입 수정 long to int. null 체크 Optional 추가 적용

+++ payment/src/main/java/com/ll/payment/model/dto/DepositInfoResponse.java
@@ -2,7 +2,7 @@ package com.ll.payment.model.dto;
 
 public record DepositInfoResponse(
         String userCode,
-        Long balance
+        Integer balance
 ) {
 }
